Add support for storing the front and back links using anydbm.
This is simplest thing style solution, to move the prototype along.

DB files are created in the instance root, one for frontlinks, one
for backlinks.

For frontlinks the key is a bag:tiddler_title pair and the value is
a list of links \0 separated. Where the links are tiddlers, they are
listed as bag:tiddler_title pairs.

For backlinks the key is a bag:tiddler_title pair and the value is a
list of bag:tiddler_title pairs that link _to_ the key.

More refinements to come.