Establish a simple handler for getting front and back links for a tiddler over HTTP
This is currently not working quite right as there are issues with
displaying non-tiddler links. Committing for sake of review.

The question at hand here is if mocking the links as tiddlers makes
is so that we can use the serialization and filtering subsystems without
causing a mess of other problems.