Correctly handle links where are just '@space'.
This had been resulting in the code exiting with an error, because
there was no where to link to. Now the data is handled and
_skipped_ until we decide where such thing should go. The obvious
place is the root of the space, but that is tiddlyspace specific
code and some determination of how to integrate that needs to be
made.