improvements in data storage and presentation
Don't use the natural key as the primary key constraint on the table,
as this leads to indexes needing to be too small, eventually leading to
integrity errors when long tiddler titles are the same for the first
few hundred characters, but not after that.

Instead have an artificial id column, and use larger indexes on the source
and target columns. This has the benefit of not limiting source and
target as much.

Also, when presenting, calculate where in a recipe the target is and
use the resulting bag for the construction of uris. This is because
when we make links (with or without @space) we are almost always linking
in the context of a recipe, not a bag.

This leaves behind some issues with links between space domains that
will 404 because of ControlView that will 200 when in the tiddlyspace.com
domain. To fix this we'll need some special presentation handling in
the tiddlyspace context.