Adjust handling of @space style links
This tries to make the tiddlyspace parts more a) correct b) easier
to factor out.

Made a fair amount of headway on a), such that frontlinks of the
@foo type will be presented, and links to spaces are handled as
fully qualified links.

However b) has not gotten any better. In fact maybe worse.

More work to be done at another time.