Fix links in the face of changes to _canonical_uri
The original links relied on behavior in the listing
of tiddlers with _canonical_uri's that has since been
removed. This results in tiddlers missing info that is
required to create URIs.

This new code gets rid of that support but in the process
has to remove some functionality: plain space links and
remote links (whether plain links or freelinks to hrefs)
are recorded but not presented in the output.

With some thought it ought to be possible to get the
functionality back, but the immediate bug needs to be handled.

In exchange for the removed support, there's now experimental
handling of parsing markdown files for links. It is quite
messy and will probably only work partially, but it is a start.
Maybe @fnd would like to make it better?